+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 -->
 
 ## 2023.10.4
+- UX: fetch account when mention component is rendering
+- UX: "missing" account now show missing image icon as avatar next to user mention
 
 ## 2023.10.3
 - UI: highlight mentions